refactor(frontend): use generated API client instead of legacy BackendAPI
Switch rate limit admin UI to use auto-generated hooks/types from
OpenAPI spec instead of manually typed legacy BackendAPI methods.
Removes UserRateLimitResponse from deprecated types.ts and
getUserRateLimit/resetUserRateLimit from legacy client.ts.
Also regenerates API client to fix pre-existing type error from dev.
