+++ autogpt_platform/frontend/src/lib/autogpt-server-api/client.ts
@@ -53,7 +53,6 @@ import type {
   TransactionHistory,
   User,
   UserPasswordCredentials,
-  UserRateLimitResponse,
   UsersBalanceHistoryResponse,
   WebSocketNotification,
 } from "./types";
@@ -479,16 +478,6 @@ export default class BackendAPI {
     transaction_filter?: string;
   }): Promise<UsersBalanceHistoryResponse> {
     return this._get("/credits/admin/users_history", params);
-  }
-
-  getUserRateLimit(user_id: string): Promise<UserRateLimitResponse> {
-    return this._get("/copilot/admin/rate_limit", { user_id });
-  }
-
-  resetUserRateLimit(user_id: string): Promise<UserRateLimitResponse> {
-    return this._request("POST", "/copilot/admin/rate_limit/reset", {
-      user_id,
-    });
   }
 
   ////////////////////////////////////////
